relocate server_config
