+++ cibo/events/input.py
@@ -10,7 +10,7 @@ from cibo.exceptions import (
     CommandUnrecognized,
     InputNotReceived,
 )
-from cibo.models.server_config import ServerConfig
+from cibo.server_config import ServerConfig
 
 
 class InputEvent(Event):
